# swagger/default.yaml
openapi: '3.0.0'
info:
  title: TLU
  description: ""
  version: "1.0.0"
paths:
  /api/v1/login:
    post:
      operationId: login
      summary: User signs in
      description: ""
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                user:
                  type: object
                  properties:
                    email:
                      type: string
                    password:
                      type: string
      responses:
        200:
          description: OK.
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                  createdAt:
                    type: string
                  _id:
                    type: string
        422:
          description: bad email or password
        401:
          description: unauthorized
  /api/v1/signup:
    post:
      operationId: signup
      summary: User creates account
      description: ""
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                user:
                  type: object
                  properties:
                    email:
                      type: string
                    password:
                      type: string
      responses:
        200:
          description: OK.
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                  createdAt:
                    type: string
                  _id:
                    type: string
        422:
          description: bad email or password
        401:
          description: unauthorized 